fixed wrong arguments on test_local_enforcer
Summary:
create_credit_update_response was used with reversed arguments. The call to this function was not failing because those arguments were casted to the specific type (bool into unit64 and uint64 into bool).

This diff reverses those arguments. Fortunately the tests seems to pass either way since all the tests were related with quota exhaustion, which meant the value was not important as long as it was lower than the used by the test.

```
void create_credit_update_response(
  const std::string& imsi,
  uint32_t charging_key,
  uint64_t volume,
  bool is_final,
  CreditUpdateResponse* response);
```

Differential Revision: D21539634

fbshipit-source-id: a6823159649a0a2b78ecbba38f00bbd6ad37df89