gc refactor (#616)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/616

Rely on single proto schema for both graph (server) and jobrunner (client)
There are two reasons why I do this change:
1. Keeping the schema files in two places is a source for problems when we change things in one schema and not in the other one
2. There is currently the following warning in tests that complain on this situation
```
2020/05/12 19:23:41 WARNING: proto: message graph.Tenant is already registered

	previously from: "github.com/facebookincubator/symphony/graph/graphgrpc"

	currently from:  "github.com/facebookincubator/symphony/jobrunner"

A future release will panic on registration conflicts. See:

https://developers.google.com/protocol-buffers/docs/reference/go/faq#namespace-conflict

2020/05/12 19:23:41 WARNING: proto: message graph.TenantList is already registered

	previously from: "github.com/facebookincubator/symphony/graph/graphgrpc"

	currently from:  "github.com/facebookincubator/symphony/jobrunner"

A future release will panic on registration conflicts. See:

https://developers.google.com/protocol-buffers/docs/reference/go/faq#namespace-conflict
```

Also pay attention that I move the schema file to a subdirectory `schema`. The reason for that is that I didn't want jobrunner to contain all ent code and more that is inside `graphgrpc` and this way it depends on a much leaner package.

Reviewed By: idoshveki

Differential Revision: D21531077

fbshipit-source-id: eb4e7eaba88369bbc14fa9680e5a48e89e4d998b