Adding Apply\Revert to user's card
Summary:
Some complained the lack if it results in a odd experience.
Plus, it gets us to problematic situation. For example, if there is a validation error in the form (i.e. missing mandatory field), other changes (i.e. role changing) doesn't get auto-saved and the user have no idea.

Reviewed By: dlvhdr

Differential Revision: D21520212

fbshipit-source-id: 79be5af7fbca52659bea1b5e5041d0c5edd21b39