Update base orchestrator plugin to use MconfigBuilder RPC
Summary:
This diff updates the base orchestrator mconfig implementation to use the
newly defined RPC interface. This is the first of all mconfig builders to have their
implementations swapped to the RPC impl as we begin to convert orc8r plugins to RPC calls.

We're not currently updating the plugin interface. This means that we have to convert between
the plugin interface parameters and the RPC interface messages. While this adds inefficiency,
it's temporary until we can directly call with gRPC.

The only net effect this change should have on behavior is that empty maps returned from `Build`
will now be nil. This is due to conversion with un/marshaling Any protos. Since this was already
being done before mconfigs were streamed to the gateways, gateways shouldn't see any differences in
their mconfigs.

Differential Revision: D21396496

fbshipit-source-id: 9a9656e5eeafb6346b7e54d034126d6da2358d68