Fix race on DB vs. state service view of registered indexers
Summary:
If state service gets indexer info for an indexer that is (no longer) registered, it can't call any of the indexer's methods (e.g. `GetVersion`), so it needs to skip reporting metrics for that indexer.

This was a racy issue in the test code since, within a single test case, indexers are being successively registered and deregistered in the service-level registry, but the DB's understanding of the indexers is not reset over the course of a single test case.

Reviewed By: uri200

Differential Revision: D21633950

fbshipit-source-id: 9b616ac682f763d00b5340a2d25ab903b940350c