Updated FEG Vagrant VM to run with dockerized FEG
Summary:
This diff updates the FEG VM, which was still using not dockerized version of FEG.

Note that we do not need this VM for development or prod. But some external contributors are not able to run FEG easily on baremetal due to missing libraries to install docker or other issues with APT.

Also, note that same behaviour may be achieved with ```install_gateway.sh```. But that script may require ubuntu 18.04. So this VM allows easy deployment of FEG

Reviewed By: emakeev

Differential Revision: D21671539

fbshipit-source-id: df6006ee6e4cd265853b2eb64d785ab84d3778dc