Refactor LTE subscriber streaming to use StreamProvider RPC implementation
Summary:
This diff converts the stream provider of the LteOrchestrator plugin to
use an implementaton of the StreamProvider RPC interface. This comes as a part of
the effort to decompose the Orchestrator plugin into strictly RPC interfaces.

Since we're leaving the existing StreamProvider plugin interface as is, for now
we're only swapping in an implementation change. To do this, we need to define
a stream provider for the LTE module that can handle multiple stream names, farming
out the stream update request to the appropriate implementor (i.e. SubscriberProvider).

To avoid import cycles we move the subscriber stream provider to the lte plugin
directory. Eventually the `legacy_stream_providers` directory will be deleted
and we will create a service out of the plugin directory
(mconfig builder, stream providers, etc.).

Differential Revision: D21670765

fbshipit-source-id: f5fc73f709066952316153d924ed981034a0a7b4