Update Orchestrator's mconfig stream to use StreamProvider RPC implementation
Summary:
This diff converts the stream provider of the BaseOrchestrator plugin to
use an implementaton of the StreamProvider RPC interface. This comes as a part of the
effort to decompose the Orchestrator's plugin into strictly RPC interfaces.

Since we're leaving the existing StreamProvider plugin interface as is, for now
we're only swapping in an implementation change. To do this, we need to define
a stream provider for a given module that can handle multiple stream names, farming
out the stream update request to the appropriate implementor (i.e. ConfigProvider).

To avoid import cycles we move the stream providers impls to the `pluginimpl`
directory. Eventually the `legacy_stream_providers` directory will be deleted
and we will create a service out of the plugin directory
(mconfig builder, stream providers, etc.).

Differential Revision: D21670764

fbshipit-source-id: 28b3966c03b100c76cfbdf65818198ad1f2f5868