Make Wallet Exhaust Behavior Configurable from Orc8r (#1717)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1717

We have a feature currently in SessionD, that terminates sessions when the subscriber's wallet is empty in CWF.
This behavior is hardcoded as a CWF only feature, and the way we detect the wallet exhaustion is also hardcoded in SessionD. (The way we support is to see how many Gx tracked rules are active)

This was fast hack to implement this feature back in Jan. For upcoming trials, we will want to support other methods of detecting wallet exhaustion. Eventually, this configuration should be exposed to the orc8r API.
For now, I'll make this behavior configurable through the Mconfig.
FWA gateways will have this feature turned off, and CWF gateways will have it turned on with the Gx Rules detection method.

Reviewed By: xjtian

Differential Revision: D21638854

fbshipit-source-id: a5d9f6eb9eabd300c0355ffeaf3aa95beeced401