Test Session IDLE timer feature (#1728)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1728

There's a gateway configuration for the AAA service that kicks off a session that has been idle for >configured amount of time.

Since the default value for idle timer is big, I'm swapping in a smaller value for this test by overwriting the default mconfig. (The value is big by default because we don't want this feature interfering with the other tests)

Also some light refactor to the fabfile to use helper functions `_switch_to_vm` & `_switch_to_vm_no_destroy`

Reviewed By: uri200

Differential Revision: D21557933

fbshipit-source-id: 276c16fe172fe6433997d38f4cf061b42bc11952