Fix standby session timeout bug for CWAG HA
Summary:
When we failover gateways, the existing sessions
should be transferred to the new active. Since the AAA server
sessions on the standby aren't clear, eventually these sessions
timeout and terminate the sessiond session. Since sessiond sessions
are shared via Redis, this ultimately causes a termination of an
active session on the active gateway.

This diff fixes this behavior by restarting the AAA server on the Disable
RPC call. Restarting the AAA server will clear standby sessions will ensuring
the service is still in an idle state.

Reviewed By: koolzz

Differential Revision: D21783169

fbshipit-source-id: 6dc5209b0e43f7117afebd8f9071ae5f9545261c