Remove SGW S1u IP address from local state
Summary:
The SGW S1u IP address is always read by the SPGW task from the local config, so
this should not be persisted to data store. Traffic tests were failing before
this change as the SPGW task was trying to read the value from data store at
init and overwriting the value from config with 0.0.0.0.

Differential Revision: D21797051

fbshipit-source-id: 7407ccca8cb114a96bc41b695554e8f7ac14d175