Remove SGW S1u IP address from local state (#1741)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1741

The SGW S1u IP address is always read by the SPGW task from the local config, so
this should not be persisted to data store. Traffic tests were failing before
this change as the SPGW task was trying to read the value from data store at
init and overwriting the value from config with 0.0.0.0.

Reviewed By: ardzoht

Differential Revision: D21797051

fbshipit-source-id: be42f9124e0955918a179715ee5ce694c1136f76