Remove EXTRA_QUOTA_MARGIN (#1712)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1712

Extra quota was added in the past to terminate the session after quota was exhausted + safety margin. The intention of this was to terminate the session if we lose connectivity with the core. The extra_quota_margin, gave provide some margin to avoid terminations due to spurial disconnections.

That feature caused side effects, plus we shouldn't be checking network issues inferring them through quota. Also note sessiond should be executing the termination, but not deciding if the the session needs to be terminated. That should be decided by core

The new behaviour is: if we go over quota, but not final unit is received, sessiond will just record the usage but not terminate.  With that we make sure sessiond do not decide over the session, and we have a simpler algorithm

Differential Revision: D21705312

fbshipit-source-id: 3c3f2f59bf6cc132a4ff9c50b03aeeec3eb12204