Fix ephemeral final action unit and info
Summary:
```is_final_grant``` was not being updated on ```update_criteria``` object when received from ```UpdateRequest```.

```final_action_info``` did not exists on ```SessionCreditUpdateCriteria``` struct, so it was never recorded

Both issues caused that after final grant was received, those attributes were not saved on the object. So next time the session was used, those settings were back to their default value. Session was never terminated by final action unit, but for quota exhaustion. When ```EXTRA_QUOTA_MARGIN``` was removed, this issue was uncovered.

This diff adds also a log line that can be use in the future to see the value of the final unit and the value of redirect server

This diff also includes a small change to help to configure cwag integ_test properly to run integ test manually from ```cwag_test```.

Differential Revision: D21822262

fbshipit-source-id: 66233d6a48872df74e5d26d76b6194c106a296a4