Serialize/deserialize session fsm state (#1761)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1761

The state was not being properly stored on stateless mode.
`TestGxUsageReportEnforcement` was not passing on stateless. This was because the termination will only complete if the session state is in a terminating state. Since the fsm_state was not set, the default value is ACTIVE -> no termination ever succeeds

modified unit test to assert serialization/deserialization happens for fsm_state

Note: I will enable statelss flag on cwf integ test to prevent these regressions in the future. (There's another bug being fixed so will wait for that)D21857619

Reviewed By: uri200

Differential Revision: D21858853

fbshipit-source-id: 984b9b1931a5cfa60636ffb8dee9f1537566cbdf