Fix version reporting for magamd docker gateways (#1763)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1763

Previously, version reporting did not work for when magmad was set to run on docker. The version would always report '0.0.0' instead of the docker image tag.
The fix was to check if the service's config init_system field is set to docker, then parse out the image tag and set the version to it.

Reviewed By: mpgermano

Differential Revision: D21844039

fbshipit-source-id: bedf90b74a793c961b3de63c345bce8c091a9e46