MME crash fix due to not sufficient queue sizes and not handling packets that could not be enqueued.
Summary:
When sctp shutdown happens, it triggers a message flooding on itti queues for main MME tasks. This results with some messages not being enqueued successfully yet the file descriptor is written to signal that a message is enqueued for the receiver end. Now, such messages are dropped and memory for them are deallocated. Error log is entered.

Additionally, the queue sizes for critical apps are increased to 2048 that should be able to handle simultaneous detach for at least that many UEs. Thus, packet drop events should be even more rare in production.

Unused task queue definitions are also removed completely for better engineering.

Differential Revision: D21892910

fbshipit-source-id: 8e1ff06680a7bc5c612e1880534bfeca4a962974