Add EventTrigger to indicate the reason for the Update Request
Summary:
Some background here:
When a session is initialized through Gx, the PCRF can install various event triggers to trigger an update request from PCEF->PCRF. The most common way is `USAGE_REPORT`, where some number of bytes is given for the session, and when the credit is used up, the PCEF is required to report back to PCRF.
However, there can be other event triggers as well, like Revalidation Timer. This is a timer set by the PCRF to trigger an update request in a set amount of time.
Each update request has to indicate what the trigger was for the request.

This diff adds an additional EventTrigger field to `UsageMonitoringUpdateRequest`
Right now, all Gx Update Requests are hardcoded to have the EventTrigger reason `USAGE_REPORT`
For other event triggers like revalidation timer, we do not want to send a usage report and the event trigger AVP needs to indicate that it was triggered by the revalidation timer.

Reviewed By: emakeev

Differential Revision: D21815637

fbshipit-source-id: 962c676836692b1985e943ff24d651fb49f5731b