Adding workflow endpoints for conductor events (#1738)
Summary:
The workflow UI has support for working with event handlers in conductor. Workflow backend endpoints are added in order for the UI to communicate with the conductor.

A workflow which has an event handler
![wflist](https://user-images.githubusercontent.com/4011326/83274358-fccecb80-a1cd-11ea-9504-b66726a86d2d.png)
The new list of event handlers for the workflow above
![eventlist](https://user-images.githubusercontent.com/4011326/83274412-1112c880-a1ce-11ea-80dd-4877f64d4130.png)
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1738

Test Plan:
Manually when creating a workflow with an event handler via builder like this
![Screenshot from 2020-05-29 17-05-44](https://user-images.githubusercontent.com/4011326/83274928-bd54af00-a1ce-11ea-9647-e4c5306a3bc8.png)

Reviewed By: Mokon

Differential Revision: D21935159

Pulled By: KenG98

fbshipit-source-id: 079d2b97178e1c8e0f9581809204524306b55f9a