Add handler to get tenant names from alertmanager-configurer (#1771)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1771

For the standalone UI we need a way to know what tenants exist in the system. In the NMS, this source of truth exists outside of configmanager. However when it's run as a standalone system, the only source for this information is the alertmanager configurer itself.

Adding this allows the UI to be populated with a list of tenants to select from

Reviewed By: xjtian

Differential Revision: D21892027

fbshipit-source-id: c742a4630e810f129a81f90d374950fe4b191e1f