Solve race condition for memory allocation from the itti memory pool to the itti messages
Summary: ITTI library utilizes a preallocated memory pool to be used for ITTI messages. As all the threads are using the same memory pool, race conditions should be handled via atomic function calls. In acquiring an index from the memory pool, the steps were not atomic, leading to the potential allocation of the same index to more than one ITTI message. This was causing issues at the time of freeing itti messages caught by assert statements. In PR#1015 in github repo, Pavel provided a solution for this which was potentially creating an infinite loop. As original PR was never modified based on the reviews, this diff is created to address this specific issue.

Differential Revision: D21989708

fbshipit-source-id: d503b135be32704a4d15940acc2ed07ce2cb14f2