Add pagination and custom search support to Log Table
Summary:
Used a custom remote data function to ensure that we can paginate.
The remote data function query also contains the search terms which we can use to filter the results with.

Note that we currently don't have support to sort the results yet. We have to add backend support for sorting the results.
Additionally elastic search 'from' clause only supports size upto 10,000, beyond that we have to use search_after clause and we need to add backend support for that as well.

Finally the export mechanism currently exports a page size worth of information which isn't very useful. Will make additional changes in upcoming reviews to ensure that we can save the query and export the entire state.

Reviewed By: Scott8440

Differential Revision: D21996655

fbshipit-source-id: e19bb83968749ecec6fbf189bc83eeb1cf64c13b