Convert metrics exporters to be called over gRPC
Summary:
As part of the transition to the orc8r service mesh, the `GetMetricsProfiles` plugin method (custom collectors and exporters) is being replaced with custom exporters accessible over RPC.

## Architecture changes

Custom collectors are being replaced in favor of opinionated metric collection.

Custom exporters are made available by services implementing the `MetricsExporter` servicer, where we'll use service labels (k8s for prod and docker-compose for dev) to provide service discovery.

In the interim, before service discovery is enabled, we retain the `MetricsProfile` concept in the plugin iface, where the registered exporters are now of type `remoteExporter`, which forward requests over RPC to the relevant service.

## Existing exporters

We have two existing exporters -- one pushing to Prometheus pushgateways, and one pushing to ODS. For the former (the standard exporter), we use it as the metricsd exporter. For the latter (FB-internal), we create the `fbinternal` service (as a precursor to the `fbinternal` service replacing the FB-internal plugin) as an attach point for the ODS exporter.

Reviewed By: xjtian

Differential Revision: D22133226

fbshipit-source-id: 5794ee53855a2165ed0cfd3bc6a9a5ba109a319c