Fix bug in state sync on SessionD restart
Summary:
We were observing on GWs with stateless sessiond that on restart, we always see `Failed to mark a scheduled dynamic rule as installed with rule_id: whitelist_sid-IMSI460660003600005`

This was occurring because traversal of `scheduled_dynamic_rules` and `dynamic_rules` were using the same variable. So first, we append dynamic_rules ids into the `rule_ids` vector. We do some logic based on that vector and then we append on scheduled dynamic rule ids onto the same vector.
This meant that the logic meant for scheduled dynamic rules was being applied to both scheduled and not scheduled dynamic rules

Reviewed By: uri200

Differential Revision: D22233904

fbshipit-source-id: 8b06bbf7f9ac450bb337ad3481e114c2969e4124