Include EventTrigger information on CreateSessionResponse even if there are no usage monitors installed
Summary:
Since EventTriggers are installed on a session level, so it should be sent back similarly to how rule installs are send down. Previously, we were handling event trigger installs inside of `UsageMonitroingUsageResponse`. This doesn't work when there are no usage monitors installed. So adding the event triggers to the top level.

```message CreateSessionResponse {
  repeated CreditUpdateResponse credits = 1; // these will only have updates
  repeated UsageMonitoringUpdateResponse usage_monitors = 6;
  repeated StaticRuleInstall static_rules = 7; // static rules
  repeated DynamicRuleInstall dynamic_rules = 8; // dynamic rules
  string session_id = 9;
  TgppContext tgpp_ctx = 10;
  repeated EventTrigger event_triggers = 11; <- NEW
  google.protobuf.Timestamp revalidation_time = 12; <- NEW
}
```

We still need to leave the event trigger related fields inside of `UsageMonitoringUsageResponse` because of the way we send `UpdateRequest`s. For updates, we have a GRPC message per monitor/charging credit update. We should migrate the UpdateRequest request to be an array of session level updates, with monitor/charging level updates nested inside. But that change is out of scope for this diff.

I've also added an additional unit test to test the case with a revalidation event trigger but no usage monitor.
I've also modified the integ test to do the same.

Reviewed By: uri200

Differential Revision: D22192106

fbshipit-source-id: 2465e6f9b2776eb876cc76c34d60ec29a33b8e92