Add go client_api for eventd
Summary:
This diff implements a golang client for the
eventd service. The client allows us to specify the event
verbosity for the given event. If the configured verbosity
is lower than the event's verbosity, no event will be sent.

This allows us to configure a verbosity such as -1 to turn
off the entire event logging system. This could be useful
if the system resources are maxing out.

Reviewed By: emakeev

Differential Revision: D21992925

fbshipit-source-id: f24efbd0a78ff3f205624819009269bba393692c