Adding scripts to configure stateless mode (#1916)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1916

This change adds scripts to enable, disable and check stateless mode for each
AGW service. The primary use of these services is for manual testing/debugging
and to use for integration testing.

In future, the config for all services will be controlled by a single flag in
magmad config.

Reviewed By: ardzoht

Differential Revision: D21924443

fbshipit-source-id: f7dbc05e813959ceeaaa2c6fc6361bd24744523a