Force restart sctpd when enabling stateless mode on AGW
Summary:
When the stateless mode flag is enabled in MME, the sctpd service does not reset
the connection with eNB. But the first time the stateless feature is enabled,
the MME has no stored state for eNB, so sctpd should restart and force eNB to
send a new S1 setup request. This change modifies the config_stateless_agw.sh
script to force restart sctpd, when switching the modes.

Reviewed By: ardzoht

Differential Revision: D22023911

fbshipit-source-id: 20fbab3c7d6863a97ba5f6a30587414f7428fe1c