Refactor SessionCredit to handle empty GSU case properly
Summary:
# A Very Long Background {emoji:1f605}
This diff tackles the specific case where we receive a GSU with all three fields, (total, rx, tx), are marked as invalid. For more context, when SessionProxy deals with the `GrantedServiceUnit` AVP, it translates the information into this protobug message.
```message CreditUnit {
  bool is_valid = 1;
  uint64 volume = 2;
}

message GrantedUnits {
  CreditUnit total = 1;
  CreditUnit tx = 2;
  CreditUnit rx = 3;
}
```
For each credit type, if an AVP is specified `is_valid` is set to `true` and the `volume` is filled in. Otherwise, `is_valid` is false and `volume` is set to 0.
The GX spec specifies in the doc below that there that there are 4 valid combinations of these grant types. (total only, rx only, tx only, and rx and tx).
(https://www.etsi.org/deliver/etsi_ts/129200_129299/129212/09.03.00_60/ts_129212v090300p.pdf) PG 29

The question here is what the intended behavior is for Gy. For example, if both total and rx are specified, do we just trigger on whichever threshold is met?
My assumption in the previous refactor was that since Gx and Gy both use the same AVP for GSU, the behavior is the same. But this could be a wrong assumption, so we can fix that if it is wrong.

I previously refactored `SessionCredit` to look at the GSU proto message, and infer how we should decide when the credit has exhausted its quota. D22115988 (https://github.com/facebookincubator/magma/commit/76017999a1ee6b45e86015856690310b801e3943). Here's some background into why I made that change.
 ---
## A bit about the implementation details pre-refactor
Before this change, when we received a volume with `is_valid=false`, we would set the volume to 0 for the following cases:

- Receiving existing / Initializing Gy ChargingCredits
- Receiving existing Gx UsageMonitors

We would set the volume to max int for
- Initializing a Gx UsageMonitor

I think the original intention here was that for Gy charging grants, treat invalid GSU as just receiving 0 credit. This way, even after receiving an invalid credit, we correctly calculate its quota status as "exhausted".
For Gx, I'm honestly not sure as to why we treat initializing and updating an existing monitor differently. One thing worth noting is that `SessionProxy` will disable any monitors with 0 GSU. So this might not matter either way. (SP code: https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/magma/feg/gateway/services/session_proxy/credit_control/gx/model_conversion.go?lines=300)

But since we don't persist which values are valid or not, this poses an issue. For example, take a case where we receive Tx=400 and Rx=500 values but an invalid Total value. (And that this credit is from Gy.) With the previous implementation, we would receive the following: (total=0, tx=400, rx=500). The following is a snippet of how we calculate whether the quota has been exhausted:
```bool SessionCredit::is_quota_exhausted(float usage_reporting_threshold) const {
  // used quota since last report
  uint64_t total_reported_usage = buckets_[REPORTED_TX] + buckets_[REPORTED_RX];
  uint64_t total_usage_since_report = std::max(uint64_t(0), buckets_[USED_TX] + buckets_[USED_RX] - total_reported_usage);
  uint64_t tx_usage_since_report = std::max(uint64_t(0), buckets_[USED_TX] - buckets_[REPORTED_TX]);
  uint64_t rx_usage_since_report = std::max(uint64_t(0), buckets_[USED_RX] - buckets_[REPORTED_RX]);

  // available quota since last report
  auto total_usage_reporting_threshold = std::max(0.0f, (buckets_[ALLOWED_TOTAL] - total_reported_usage) * usage_reporting_threshold);

  // reported tx/rx could be greater than allowed tx/rx
  // because some OCS/PCRF might not track tx/rx,
  // and 0 is added to the allowed credit when an credit update is received
  auto tx_usage_reporting_threshold = std::max(0.0f, (buckets_[ALLOWED_TX] - buckets_[REPORTED_TX]) *  usage_reporting_threshold);
  auto rx_usage_reporting_threshold = std::max(0.0f, (buckets_[ALLOWED_RX] - buckets_[REPORTED_RX]) * usage_reporting_threshold);

  if (total_usage_since_report >= total_usage_reporting_threshold) {
    return true;
  } else if ((buckets_[ALLOWED_TX] > 0) && (tx_usage_since_report >= tx_usage_reporting_threshold)) {
    return true;
  } else if ((buckets_[ALLOWED_RX] > 0) && (rx_usage_since_report >= rx_usage_reporting_threshold)) {
    return true;
  }
  return false;
}
```
When we check `total_usage_since_report >= total_usage_reporting_threshold`, we incorrectly infer that quota has been exhausted here.

 ---
# What this diff changes
The refactor linked above had a bug where on receiving invalid GSUs the grant_tracking_type was set to `NO_TRACKING`. This was actually wrong because, we could have a case where we don't receive a valid GSU but receive a valid FinalAction. So to fix this, on receiving an invalid GSU, we should just leave the TrackingType as it is. This way, even after receiving an invalid GSU, if the quota was exhausted before, it will continue to be exhausted.
This change means that the type `NO_TRACKING` is no longer needed, so we'll just get rid of it here.
Additionally, now that we have a functionality to persist grant tracking type and not calculate it everytime  we receive a credit, we will initialize the tracking type to `TOTAL_ONLY` on `SessionCredit` creation.

I've added unit tests on the credit level and session state level to cover these cases.

Reviewed By: uri200

Differential Revision: D22389654

fbshipit-source-id: fc4a8b16fafb79c1ee95391959da8a490241df87