Add change to not create a monitor when there is not monitor given
Summary:
We currently overload `UsageMonitoringUpdateResponse` with non-usage monitor related information as well.
```message UsageMonitoringUpdateResponse {
  UsageMonitoringCredit credit = 1;
  string session_id = 2;
  string sid = 3; // System Identification Number
  bool success = 4;
  repeated EventTrigger event_triggers = 5; <- THIS
  google.protobuf.Timestamp revalidation_time = 6; <- THIS
  uint32 result_code = 7;
  repeated string rules_to_remove = 8; <- THIS
  repeated StaticRuleInstall static_rules_to_install = 9; <- THIS
  repeated DynamicRuleInstall dynamic_rules_to_install = 10; <- THIS
  TgppContext tgpp_ctx = 11; <- THIS
}
```
This means that we could make an update exchange without any usage monitor involved. Say an update request is sent up due to a revalidation timer, then we would most likely receive a response with an empty UsageMonitoringCredit.

Tldr is that we should check whether the credit field is filled in before trying to create a monitor out of it. :)

Reviewed By: uri200

Differential Revision: D22390759

fbshipit-source-id: 0aef818d885852e5fd35092f326e04d405f78008