Convert application Terraform to single-apply
Summary:
This diff reduces the number of required `terraform apply` calls from 5 to 3.

(Actually, it also allows moving down to 2, but I think it's less error-prone to leave it at 3, treating secrets upload as a separate step.)

See T69025136 for additional information.

Previously, we performed the application Terraforming in three steps:

- Install orc8r, without NMS
- Create admin user in certifier, then upload those secretes to AWS Secrets Manager
- Install orc8r, with NMS

Now, instead, we generate the admin user certs ahead of time, allowing us to just
- Install orc8r, with NMS
- Add existing admin user in certifier

Reviewed By: xjtian

Differential Revision: D22273295

fbshipit-source-id: b936e6421e2562ec8715c6a7078700a4222a9901